+++ assertr/src/assertions/core/option.rs
@@ -1,15 +1,19 @@
 use crate::{AssertThat, Mode, actual::Actual, mode::Panic, tracking::AssertionTracking};
-use core::fmt::Debug;
+use alloc::string::String;
+use core::fmt::{Debug, Write};
 use core::option::Option;
+use indoc::writedoc;
 
 /// Data-extracting assertion for `Option` values.
-/// Only available in Panic mode, as the extracted `T` cannot be produced when the value is `None`.
 #[cfg_attr(feature = "fluent", assertr_derive::fluent_aliases)]
 pub trait OptionExtractAssertions<'t, T> {
     /// Test if this option is of the `Some` variant.
     /// This is a terminal operation on the contained `Option`,
     /// as there is nothing meaningful to do with the option if its variant was ensured.
     /// This allows you to chain additional expectations on the contained success value.
+    ///
+    /// Only available in `Panic` mode, as the extracted `T` cannot be produced when the value is
+    /// `None`. Use `OptionAssertions::is_some_satisfying` for capture mode.
     fn is_some(self) -> AssertThat<'t, T, Panic>
     where
         T: Debug;
@@ -24,10 +28,14 @@ impl<'t, T> OptionExtractAssertions<'t, T> for AssertThat<'t, Option<T>, Panic> 
         self.track_assertion();
 
         if !self.actual().is_some() {
-            self.fail(format_args!(
-                "Actual: {actual:#?}\n\nis not of expected variant: Option:Some\n",
-                actual = self.actual()
-            ));
+            let actual = self.actual();
+            self.fail(|w: &mut String| {
+                writedoc! {w, r"
+                    Actual: {actual:#?}
+
+                    is not of expected variant: Option::Some
+                "}
+            });
         }
 
         self.map(|actual| match actual {
@@ -39,8 +47,15 @@ impl<'t, T> OptionExtractAssertions<'t, T> for AssertThat<'t, Option<T>, Panic> 
 
 /// Non-extracting assertions for `Option` values.
 /// These work in any mode (Panic or Capture).
+#[allow(clippy::return_self_not_must_use)]
 #[cfg_attr(feature = "fluent", assertr_derive::fluent_aliases)]
 pub trait OptionAssertions<'t, T, M: Mode> {
+    /// Test if this option is of the `Some` variant, then run additional assertions on the contained value.
+    fn is_some_satisfying<A>(self, assertions: A) -> Self
+    where
+        T: Debug,
+        A: for<'a> FnOnce(AssertThat<'a, &'a T, M>);
+
     /// Test if this option is of the `None` variant.
     /// This is a terminal operation on the contained `Option`,
     /// as there is nothing meaningful to do with the option after its variant was ensured.
@@ -51,6 +66,29 @@ pub trait OptionAssertions<'t, T, M: Mode> {
 
 impl<'t, T, M: Mode> OptionAssertions<'t, T, M> for AssertThat<'t, Option<T>, M> {
     #[track_caller]
+    fn is_some_satisfying<A>(self, assertions: A) -> Self
+    where
+        T: Debug,
+        A: for<'a> FnOnce(AssertThat<'a, &'a T, M>),
+    {
+        self.track_assertion();
+
+        if self.actual().is_some() {
+            self.satisfies_ref(|it| it.as_ref().unwrap(), assertions)
+        } else {
+            let actual = self.actual();
+            self.fail(|w: &mut String| {
+                writedoc! {w, r"
+                    Actual: {actual:#?}
+
+                    is not of expected variant: Option::Some
+                "}
+            });
+            self
+        }
+    }
+
+    #[track_caller]
     fn is_none(self) -> AssertThat<'t, (), M>
     where
         T: Debug,
@@ -58,10 +96,14 @@ impl<'t, T, M: Mode> OptionAssertions<'t, T, M> for AssertThat<'t, Option<T>, M>
         self.track_assertion();
 
         if !self.actual().is_none() {
-            self.fail(format_args!(
-                "Actual: {actual:#?}\n\nis not of expected variant: Option:None\n",
-                actual = self.actual()
-            ));
+            let actual = self.actual();
+            self.fail(|w: &mut String| {
+                writedoc! {w, r"
+                    Actual: {actual:#?}
+
+                    is not of expected variant: Option::None
+                "}
+            });
         }
 
         self.map(|_actual| Actual::Owned(()))
@@ -93,7 +135,74 @@ mod tests {
                 -------- assertr --------
                 Actual: None
 
-                is not of expected variant: Option:Some
+                is not of expected variant: Option::Some
+                -------- assertr --------
+            "});
+        }
+    }
+
+    mod is_some_satisfying {
+        use crate::prelude::*;
+        use indoc::formatdoc;
+
+        #[test]
+        fn succeeds_when_some_and_assertions_pass() {
+            assert_that!(Option::<i32>::Some(42)).is_some_satisfying(|some| {
+                some.is_equal_to(&42);
+            });
+        }
+
+        #[test]
+        fn captures_inner_failure_when_some_and_assertion_fails() {
+            let failures = assert_that!(Option::<i32>::Some(42))
+                .with_capture()
+                .with_location(false)
+                .is_some_satisfying(|some| {
+                    some.is_greater_than(&9000);
+                })
+                .capture_failures();
+
+            assert_that!(failures).contains_exactly::<String>([formatdoc! {"
+                -------- assertr --------
+                Actual: 42
+
+                is not greater than
+
+                Expected: 9000
+                -------- assertr --------
+            "}]);
+        }
+
+        #[test]
+        fn captures_variant_failure_when_none() {
+            let failures = assert_that!(Option::<i32>::None)
+                .with_capture()
+                .with_location(false)
+                .is_some_satisfying(|_| panic!("assertions should not run"))
+                .capture_failures();
+
+            assert_that!(failures).contains_exactly::<String>([formatdoc! {"
+                -------- assertr --------
+                Actual: None
+
+                is not of expected variant: Option::Some
+                -------- assertr --------
+            "}]);
+        }
+
+        #[test]
+        fn panics_when_none() {
+            assert_that_panic_by(|| {
+                assert_that!(Option::<i32>::None)
+                    .with_location(false)
+                    .is_some_satisfying(|_| panic!("assertions should not run"))
+            })
+            .has_type::<String>()
+            .is_equal_to(formatdoc! {"
+                -------- assertr --------
+                Actual: None
+
+                is not of expected variant: Option::Some
                 -------- assertr --------
             "});
         }
@@ -123,7 +232,7 @@ mod tests {
                     42,
                 )
 
-                is not of expected variant: Option:None
+                is not of expected variant: Option::None
                 -------- assertr --------
             "});
         }